proxied_controllers: split into proxied_controllers and build_proxied_controllers
